add alert confirm to delete task

+++ client/src/Components/Tasks/Task.jsx
@@ -1,29 +1,43 @@
 import { useContext } from "react";
 
+import swal from "sweetalert";
+
 import ProjectContext from "../../Context/Project/context.js";
 import TaskContext from "../../Context/Task/context.js";
 
 function Task({ task }) {
   //STATE GLOBAL PROJECT
   const globalState = useContext(ProjectContext);
-  const { projectselected} = globalState;
+  const { projectselected } = globalState;
 
   //STATE GLOBAL TASK
   const taskState = useContext(TaskContext);
   const { deleteTask, changeFormToEditTask, selectTask } = taskState;
 
   //FUNCTION TO DELETE TASK
-  const handleDeleteTask = (e)=>{
+  const handleDeleteTask = (e) => {
     e.preventDefault();
-    deleteTask(task._id)
-    
-  }
-  const handleEditTask = (e)=>{
+    swal({
+      title: `Are you sure delete ${task.name} task?`,
+      icon: "warning",
+      buttons: true,
+      dangerMode: true,
+    }).then((resOk) => {
+      if (resOk) {
+        swal({
+          title: "Task deleted",
+          icon: "success",
+        });
+        deleteTask(task._id);
+      }
+    });
+  };
+  const handleEditTask = (e) => {
     e.preventDefault();
     changeFormToEditTask();
-    selectTask(task)
-  }
-  
+    selectTask(task);
+  };
+
   // const handleChangeStatus = (task)=>{
   //   // e.preventDefault();
   //   if(task.status){
@@ -39,33 +53,37 @@ function Task({ task }) {
       <p>{task.name}</p>
       <div className="estado">
         {task.status ? (
-          <button type="button" className="completo"
-          //  onClick={()=>handleChangeStatus(task)}
-           >
+          <button
+            type="button"
+            className="completo"
+            //  onClick={()=>handleChangeStatus(task)}
+          >
             Complete
           </button>
         ) : (
-          <button type="button" className="incompleto" 
-          // onClick={()=>handleChangeStatus(task)}
+          <button
+            type="button"
+            className="incompleto"
+            // onClick={()=>handleChangeStatus(task)}
           >
             Incomplete
           </button>
         )}
       </div>
-      
+
       <div className="acciones">
-        <button 
-        type="button" 
-        className="btn btn-primario"
-        onClick={handleEditTask}
+        <button
+          type="button"
+          className="btn btn-primario"
+          onClick={handleEditTask}
         >
           Edit
         </button>
 
-        <button 
-        type="button" 
-        className="btn btn-secundario"
-        onClick={handleDeleteTask}
+        <button
+          type="button"
+          className="btn btn-secundario"
+          onClick={handleDeleteTask}
         >
           Delete
         </button>
